refactor: rename GiveItemTypeEvent >>> PurchaseItemEvent
